🐛 FIX: 버그 수정
tinyint는 postgres에서 지원하지 않기때문에 smallint로 변경

 #5

+++ src/chat-room/entities/chat-room.entity.ts
@@ -80,7 +80,7 @@ export class ChatRoom {
     updatedAt:Date;
 
     // 활성화 여부 
-    @Column({type:'tinyint',default:1,name:'is_active'})
+    @Column({type:'smallint',default:1,name:'is_active'})
     isActive:number
 
     // 메세지와 chatRoom 조인
